+++ docs/versionhistory.rst
@@ -19,6 +19,8 @@ This library adheres to `Semantic Versioning 2.0 <https://semver.org/#semantic-v
   sequence, mapping) to only check the first item by default. To get the old behavior,
   set ``typeguard.config.collection_check_strategy`` to
   ``CollectionCheckStrategy.ALL_ITEMS``
+- **BACKWARD INCOMPATIBLE** Type checking failures now raise
+  ``typeguard.TypeCheckError`` instead of ``TypeError``
 - Dropped Python 3.5 and 3.6 support
 - Dropped the deprecated profiler hook (``TypeChecker``)
 - Added a configuration system