+++ docs/versionhistory.rst
@@ -3,7 +3,7 @@ Version history
 
 This library adheres to `Semantic Versioning 2.0 <https://semver.org/#semantic-versioning-200>`_.
 
-**UNRELEASED**
+**3.0.1** (2023-03-16)
 
 - Improved the documentation
 - Fixed assignment unpacking (``a, b = ...``) being checked incorrectly